return grp array as integer array to avoid issues with things like bridge

+++ pynbody/halo.py
@@ -1014,7 +1014,7 @@ class AHFCatalogue(HaloCatalogue):
 
                 ar[id_t] = hcnt[cnt]
             cnt += 1
-        return ar
+        return ar.astype(np.int)
 
     def _setup_children(self):
         """
